completed the command node execution
