+++ inc/data_structs.h
@@ -1,17 +1,17 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ast.h                                              :+:      :+:    :+:   */
+/*   data_structs.h                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: zyeoh <zyeoh@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/09 13:29:22 by zyeoh             #+#    #+#             */
-/*   Updated: 2024/04/10 14:59:21 by zyeoh            ###   ########.fr       */
+/*   Updated: 2024/04/10 16:27:57 by zyeoh            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#ifndef AST_H
-# define AST_H
+#ifndef DATA_STRUCTS_H
+# define DATA_STRUCTS_H
 
 # include "minishell.h"
 
@@ -25,7 +25,7 @@ typedef enum e_nodetype
 typedef struct s_node
 {
 	t_nodetype				type;
-	union							//  * Do not typedef union, will result in node->union_name->var
+	union //  * Do not typedef union, will result in node->union_name->var
 	{
 		struct s_pipe		*pipe;
 		struct s_command	*command;
@@ -38,11 +38,11 @@ typedef struct s_command
 	char					*str;
 }							t_command;
 
-// typedef struct s_redir
-// {
-// 	t_node			*input;
-// 	t_node			*output;
-// }					t_command;
+typedef struct s_redir_in
+{
+	t_node			*input;
+	t_node			*output;
+}					t_redir_in;
 
 typedef struct s_pipe
 {
