+++ inc/minishell.h
@@ -6,17 +6,22 @@
 /*   By: zyeoh <zyeoh@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/08 17:04:32 by zyeoh             #+#    #+#             */
-/*   Updated: 2024/04/09 19:05:55 by zyeoh            ###   ########.fr       */
+/*   Updated: 2024/04/10 16:26:55 by zyeoh            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef MINISHELL_H
 # define MINISHELL_H
 # include "../libft/libft.h"
-# include "ast.h"
+# include "data_structs.h"
 # include <readline/history.h>
 # include <readline/readline.h>
 # include <stdio.h>
 # include <sys/wait.h>
+
+// execute.c
+void	execute(t_node *node);
+char	*get_exec(char **cmd_path, char *cmd);
+void	ex_cmd(t_command *command);
 
 #endif
