--- inc/temp.h
@@ -1,28 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   temp.h                                             :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: zyeoh <zyeoh@student.42.fr>                +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/04/09 13:29:22 by zyeoh             #+#    #+#             */
-/*   Updated: 2024/04/09 14:27:13 by zyeoh            ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#ifndef TEMP_H
-# define TEMP_H
-
-typedef struct s_compound_word
-{
-	char	*str;
-}			t_compound_word;
-
-typedef struct s_and_or
-{
-	int		id;
-    
-
-}			t_and_or;
-
-#endif
